requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.50.0
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.16.0
zipp==3.18.1
